Bug fix in testnet chain validation
I'm a bit stumped by the testnet chain validation rules. The bitcoin core code
clearly does not allow a minimum difficulty block to be mined at an epoch (difficulty
adjustment block) but yet it regularly happens on the testnet. Since this code
mirrors bitcoin core, the chain validation fails when it hits such a block. For
this reason I've disabled checking the difficulty of blocks when on testnet.
If I figure out what the issue is I can re-enable it, but for our purposes, we
just need a working testnet to make transactions.