Add recover() to startChainDownload
There's likely a race condition that is occasionally causing the blockQueue to
be closed more than once resulting in an unhandled panic. This commit adds
a recover to the function to prevent the app from crashing but more investigation
is need to determine why close is called more than once.