Refactor transaction download queues
Previously transaction hashes were held in memory along with the tx's height. The
problem was we could not delete the tx from memory after it was downloaded since
this would causes the tx's height to be overridden if a peer sent us the same tx
after the fact. Unlikely to happen, but it still prevented us from clearing out
memory.

This commit refactors the transaction download tracking so that each peer has its
own download queue. It also tracks ingested transactions in memory and the
ingest() function will discard tranasctions that have already been ingested. This
allows up to free up memory of downloaded transactions.