Refactor keys db
Currently the keys db stores the bip32 key path with the scriptPubKey of the
corresponding address. With segwit nearing activation we'll want to make sure
we can detect payments to both segwit and legacy addresses when syncing the
chain. This commit refactors the keys database to store the ScriptAddress(hash160)
of the key instead of a scriptPubKey. Since segwit addresses use the same hash160
we can fetch a key from the database given either a segwit or legacy utxo.