Improve blockchain handling
This commit adds an in-memory cache to the blockchain db to improve the speed of
lookups when validating headers. The cache will also allow for calculating median
time past needed for future extensions.

It also fixes a bug in the common ancestor reorg calculation and improves the
reorg handling.

Finally, when switching download peers and checking for more blocks it waits
until it has > half the target outgoing peers before continuing.